Refactor WebController
