+++ src/main/java/se/iths/springbootgroupproject/SpringBootGroupProjectApplication.java
@@ -2,12 +2,7 @@ package se.iths.springbootgroupproject;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import se.iths.springbootgroupproject.config.SpringSecurityAuditorAware;
-
-import java.util.Optional;
 
 @SpringBootApplication
 @EnableJpaAuditing(auditorAwareRef = "auditorAware")
