+++ src/main/java/se/iths/springbootgroupproject/config/AuditingConfig.java
@@ -4,6 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import se.iths.springbootgroupproject.audit.SpringSecurityAuditorAware;
 
 @Configuration
 @EnableJpaAuditing(auditorAwareRef = "auditorAware")
