+++ src/main/resources/templates/messages.html
@@ -46,8 +46,6 @@
             </td>
             <td th:text="${message.createdDate}">2020-01-01</td>
             <td>
-                <span th:if="${message.lastModifiedDate != null}"
-                      th:text="${message.lastModifiedDate} + ' ' + #{editedby} + ': ' + ${message.lastModifiedBy}"></span>
                 <span th:if="${message.lastModifiedBy != null}" th:text="${message.lastModifiedDate} + ' ' + #{editedby} + ': ' + ${message.lastModifiedBy}"></span>
                 <p id="edit" th:if="${message.user?.id} == ${loggedInUser?.id}">
                     <a th:href="@{/web/update/{messageId}(messageId=${message.id})}" th:text="#{editmessage}"></a>
